Switch to prepend instead of alias chaining methods
